EmulatorPkg: add emulated RpmcLib and VariableKeyLib
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=2594

With the emulated RpmcLib and VariableKeyLib, we can enable protected
variable service in emulator platform for try-run or test.

PROTECTED_VARIABLE_ENABLE is used to enable/disable the protected variable
services. Its default is FALSE and can be changed to TRUE in dsc file or
command line (-DPROTECTED_VARIABLE_ENABLE=TRUE).

Cc: Jiewen Yao <jiewen.yao@intel.com>
Cc: Nishant C Mistry <nishant.c.mistry@intel.com>
Cc: Michael Kubacki <michael.kubacki@outlook.com>
Cc: Jordan Justen <jordan.l.justen@intel.com>
Cc: Andrew Fish <afish@apple.com>
Cc: Ray Ni <ray.ni@intel.com>
Signed-off-by: Jian J Wang <jian.j.wang@intel.com>